Extracted `StubNetworkClient` structure in the separate file
